separate out list of remotion packages
