+++ packages/create-video/src/list-of-remotion-packages.ts
@@ -1,0 +1,22 @@
+export const listOfRemotionPackages = [
+	'@remotion/bundler',
+	'@remotion/cli',
+	'@remotion/eslint-config',
+	'@remotion/eslint-plugin',
+	'@remotion/renderer',
+	'@remotion/skia',
+	'@remotion/lottie',
+	'@remotion/media-utils',
+	'@remotion/motion-blur',
+	'@remotion/google-fonts',
+	'@remotion/noise',
+	'@remotion/paths',
+	'@remotion/babel-loader',
+	'@remotion/lambda',
+	'@remotion/player',
+	'@remotion/preload',
+	'@remotion/three',
+	'@remotion/gif',
+	'@remotion/shapes',
+	'remotion',
+];
